# Increase AGR performance