The current individual keyframe adding is very slow when there are a lot of keyframes.

With these changes they are all read into memory and added at once using Blender's fast native methods.

However, setting the interpolation mode is not possible using fast native methods, so I also made it selectable. If the user selects the Blender's default interpolation mode (Bezier), we don't need to implicitly set the interpolation mode, which makes the import even faster.

I also implemented the shortest rotation path taking for bones too, since the lack of it might become an issue with interpolation enabled.

Here are some import time comparisons:

AGR Size | Current release | Fork, default interpolation | Fork, Bezier interpolation
---------- | ----------------- | ---------------------------- | ------------------------------
20 MB     | 103.8334 s         | 82.5832 s                           | 49.0030 s
80 MB     | 8879.9433 s       | 1332.9248 s                       | 424.8857 s