@dtugend 2020-05-12
Thank you for your contribution. I will try to review it around the weekend.

Are you sure it's a good idea to make it take shortest path for bone rotations?
I somehow remember it not being a good one:
The game internally uses matrices for the bones, which are only parsed out as position + rotation quaternion, which is also why info that can only by expressed with matrices is lost (e.g. why cl_righthand 0 doesn't work, since that's mirroring / sort of scaling).
If I remember correctly even with correct quaternion interpolation (you can test it by enabling "Add interpolated keyframes" option, which interpolates them as new keyframes if you use higher / wrong FPS) that would lead to problems.