@lasa01 2020-05-12
I will test the interpolation more, I haven't run into any issues (at least
yet).

The keyframe_points.foreach_set method requires flattened arrays, so that
code is correct as it is.