@lasa01 2020-05-12
Reverted the shortest path commit locally to compare results.

It doesn't seem to create any issues, but it fixes the existing interpolation issues, as seen here:
![image](https://user-images.githubusercontent.com/3680681/81690938-b6a30980-9464-11ea-93e2-91bf1a610e24.png)
(I did not modify the existing implementations for the "no shortest path" image, that issue is caused by bone rotations.)

With the shortest path commit, the rotation interpolations look correct on the whole scene.