@dtugend 2020-05-12
> The keyframe_points.foreach_set method requires flattened arrays, so that
code is correct as it is.

Thanks, didn't know that, sorry.

> It doesn't seem to create any issues, but it fixes the existing interpolation issues, as seen here:
Thanks for testing.

(I am not 100% sure about this to be honest, just thinking I remember s.th. from past tests.)
It might dampen the absolute error for the wrong non-spherical Bezier interpolation, I will need to test with interpolate keyframes set and having it interpolate some to know for sure (unless you alerady did that). Will let you know by Friday hopefully, or maybe earlier, depends, a bit busy right now with other things sadly.