@dtugend 2020-05-13
Thanks for helping. You were right about all things. Sorry.